feat(ts): messari action provider refinements (#619)
